refactor embeddings files
