+++ app/packages/embeddings/src/getPointIndex.tsx
@@ -1,0 +1,12 @@
+import { findIndexByKeyValue } from "./findIndexByKeyValue";
+
+export function getPointIndex(trace, id) {
+  let idx = findIndexByKeyValue(trace, "id", id);
+  if (idx === undefined || idx === null) {
+    idx = findIndexByKeyValue(trace, "sample_id", id);
+  }
+  if (idx === null) {
+    debugger;
+  }
+  return idx;
+}
