+++ app/packages/embeddings/src/handleInitialPlotLoad.tsx
@@ -1,0 +1,16 @@
+import { getFetchFunction } from "@fiftyone/utilities";
+
+export async function handleInitialPlotLoad({
+  datasetName,
+  brainKey,
+  view,
+  labelField,
+}) {
+  const res = await getFetchFunction()("POST", "/embeddings/plot", {
+    datasetName,
+    brainKey,
+    view,
+    labelField,
+  });
+  return res;
+}
