+++ app/packages/embeddings/src/sortStringsAlphabetically.tsx
@@ -1,0 +1,5 @@
+// a function that sorts strings alphabetically
+// but puts numbers at the end
+export function sortStringsAlphabetically(a, b) {
+  return a.localeCompare(b);
+}
