+++ app/packages/embeddings/src/useKeyDown.tsx
@@ -1,0 +1,16 @@
+import { useEffect } from "react";
+
+// a react hook that returns true when the given key is down
+export function useKeyDown(key, handler) {
+  useEffect(() => {
+    const handleKeyDown = (e) => {
+      if (e.key === key) {
+        handler(true);
+      }
+    };
+    document.addEventListener("keydown", handleKeyDown);
+    return () => {
+      document.removeEventListener("keydown", handleKeyDown);
+    };
+  }, [key]);
+}
